Add files via upload
Aim: Also known as the Collatz or Syracuse hypothesis.
Input any positive random number. If it is even, it will be halved, if it is odd, it will be multiplied by 3, and 1 added.
The sequence will always end up at 1. Upper limit is defined as 100 to save programming time, but this constant can be changed.
Computer gives how many steps it took to get to 1.