Add files via upload
The ancient game of Nim is said to have originated in China—it closely resembles the Chinese game of 捡石子 jiǎn-shízi, or "picking stones".  The origin is uncertain, but there are references to it from the beginning of the 16th century.

Aim: there are 20 stones and each player will pick either 1 or 2 stones alternately. Winner does not pick last stone.
To make it more interesting, you are player 1, computer is player 2, but you choose on behalf of the computer.